Day 15 pt 2 more WIP
Having trouble setting the label on new `Lens` structs. The label is a
String returned by `parse_step`, and `Lens` needs to take a `&str` of it
when setting the label property, in order for `Lens` to implement the
`Copy` trait so that it can be used to initialise the `lenses` = `[Lens;
5]` array in the `Box` struct.

Next I'd like to try making `parse_step` return a string slice (`&str`)
instead of a `String` for the label.