Fix Day 14 pt 2
Not sure why this started happening, but the `repeating_diff` started
being set to None which then caused the `unwrap()` on it to panic.

There are values in `previously_seen` that will have None diffs - those
for patterns only seen in the first stage of cycles before a repetitive
pattern begins.

The code change means that `repeating_diff` will always be set to a Some
value.