Day 20 pt 1 (remove lifetimes, Conjunction modules not working)
Conjunction boxes need their input_module_memory initialised properly.
At the moment an input_module only gets added to the HashMap when a
Pulse is first received from it. Need to go through all of the input
after first initialising the modules and add a memory K,V for any
modules that reference the Conjunction module in question.